Generate cross-validation test-training sets manually

Generate cross-validation test-training sets from manually specified indices of observations in the test and and training sets.

crossv_df(data, ...)

# S3 method for data.frame
crossv_df(data, train = NULL, test = NULL, ...)

# S3 method for grouped_df
crossv_df(data, train = NULL, test = NULL, ...)

Arguments

data

A data frame

...

Arguments passed to methods

train

A list of integer vectors or an integer vector, with the indices of the training set. If train is NULL, then the training set is the complement of the test set.

test

A list of integer vectors or an integer vector, with the indices of the test set. If test is NULL, then the test set is the complement of the training set.

Methods (by class)

data.frame: Partition a data frame into test and training sets by row.

grouped_df: Partition a grouped data frame into test and training sets by group.

Examples

library("dplyr")

# test-training pairs can be specified by test sets
crossv_df(mtcars, train = list(1:5, 6:10))

#> # A tibble: 2 × 3
#>            train           test   .id
#>           <list>         <list> <int>
#> 1 <S3: resample> <S3: resample>     1
#> 2 <S3: resample> <S3: resample>     2

# or by training sets
crossv_df(mtcars, train = list(1:20, 10:32))

#> # A tibble: 2 × 3
#>            train           test   .id
#>           <list>         <list> <int>
#> 1 <S3: resample> <S3: resample>     1
#> 2 <S3: resample> <S3: resample>     2

# or by both (and they don't need to partition the observations)
crossv_df(mtcars,
          train = list(1:20, 10:32),
          test = list(21:32, 1:9))

#> # A tibble: 2 × 3
#>            train           test   .id
#>           <list>         <list> <int>
#> 1 <S3: resample> <S3: resample>     1
#> 2 <S3: resample> <S3: resample>     2

# with grouped data frames, the indices refer to groups
crossv_df(group_by(mtcars, mpg), train = list(1L))$train

#> [[1]]
#> <resample [2 x 11]> 15, 16
#>