Generate cross-validated leave-p-out test/training pairs

Generate cross-validated leave-one-out or leave-p-out test/training pairs. The function leave-p-out generates leave-p-out test/training pairs, while the function leave-one-out convenience function for the common case of leave-one-out cross-validation, p = 1.

crossv_lpo(data, p, ...)

# S3 method for data.frame
crossv_lpo(data, p = 1L, ...)

# S3 method for grouped_df
crossv_lpo(data, p = 1L, ...)

crossv_loo(data)

Arguments

data

A data frame or vector

p

The number of elements to include in the test set.

...

Passed to methods

Value

A data frame with \({n}\choose{p}\) rows and the following columns:

train

A list of resample objects. Training sets.

test

A list of resample objects. Test sets.

.id

An integer vector of identifiers

Methods (by class)

data.frame: Generate test/train sets by leaving rows of the data frame out.

grouped_df: Generate test/train sets by leaving groups out.

References

Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.

Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.

Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.

Examples

# Example originally from modelr::crossv_mc
library("purrr")
library("dplyr")

# LOO cross-validation
cv1 <- crossv_loo(mtcars)
models <- map(cv1$train, ~ lm(mpg ~ wt, data = .))
summary(map2_dbl(models, cv1$test, modelr::rmse))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> 0.05235 1.03900 2.30100 2.51700 3.39800 7.35900 

# Leave-p-Out cross-validation with p = 2
cv2 <- crossv_lpo(mtcars, p = 2)
models <- map(cv2$train, ~ lm(mpg ~ wt, data = .))
summary(map2_dbl(models, cv2$test, modelr::rmse))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.1205  1.7870  2.5780  2.8240  3.5060  7.9080