Generate cross-validated time-series test/train sets

Generate test/train set for time series. Each training set consists of observations before the test set. This is also called "evaluation on a rolling forecasting origin".

crossv_ts(data, ...)

# S3 method for data.frame
crossv_ts(data, horizon = 1L, test_size = 1L,
  test_partial = FALSE, train_partial = TRUE, train_size = n,
  test_start = NULL, from = 1L, to = n, by = 1L, ...)

# S3 method for grouped_df
crossv_ts(data, horizon = 1L, test_size = 1L,
  test_partial = FALSE, train_partial = TRUE, train_size = n,
  test_start = NULL, from = 1L, to = n, by = 1L, ...)

Arguments

data

A data frame

...

Arguments passed to methods

horizon

Difference between the first test set observation and the last training set observation

test_size

Size of the test set

test_partial

If TRUE, then allows for partial test sets. If FALSE, does not allow for partial test sets. If a number, then it is the minimum allowable size of a test set.

train_partial

Same as test_partial, but for the training set.

train_size

The maximum size of the training set. This allows for a rolling window training set, instead of using all obsservations from the start of the time series.

test_start, from, to, by

An integer vector of the starting index values of the test set. NULL, then these are generated from seq(from, to, by).

Value

A data frame with k the following columns:

sample

A list of resample objects. Training sets.

.id

An integer vector of identifiers.

Details

In time-series cross-validation the training set only uses observations that are prior to the test set. Suppose the time series has \(n\) observations, the training set has a maximum size of \(r \leq n\) and minimum size of \(s \geq r\). and the test set has a maximum size of \(p \leq n\) and minimum size of \(q \geq p\). For indices \(i \in \{1, \dots, N\}\):

Select observations \(i, \dots, \max{p, n}\) for the test set.

Select observations \(\max{i - h - p}, \dots, i - h\) for the training set.

If the test set has a size of at least q and the training set has a size of at least r.

Methods (by class)

data.frame: Data frame method. This rows are assumed to be ordered.

grouped_df: Grouped data frame method. The groups are assumed to be ordered, and the cross validation works on groups rather than rows.

References

Hyndman RJ (2017). forecast: Forecasting functions for time series and linear models. R package version 8.0, http://github.com/robjhyndman/forecast.

Hyndman RJ and Khandakar Y (2008). "Automatic time series forecasting: the forecast package for R." Journal of Statistical Software. http://www.jstatsoft.org/article/view/v027i03.

Rob J. Hyndman. Cross-validation for time series. December 5, 2016.

Rob J. Hyndman. Time series cross-validation: an R example. August 26, 2011.

Rob J. Hyndman and George Athanasopoulos. "Evaluating Forecast Accuracy." https://www.otexts.org/fpp/2/5.

Max Kuhn. "Data splitting for Time Series." The caret Package. 2016-11-29. https://topepo.github.io/caret/data-splitting.html.