Generate cross-validated time-series K-fold test-training pairs

Splits data into test/train sets for cross-validating time series data. In each split, the indices are increasing, so only previous data are used to test future data.

crossv_tskfold(data, K, ...)

# S3 method for data.frame
crossv_tskfold(data, K = 5L, ...)

# S3 method for grouped_df
crossv_tskfold(data, K = 5L, ...)

Arguments

data

A data frame or vector

K

The number of folds

...

Passed to methods

Value

A data frame with K the following columns:

sample

A list of resample objects. Training sets.

.id

An integer vector of identifiers.

Details

The data are split into \(K \geq 2\) folds, where for each \(i \in 2:K\), the \(i\)-th fold is the test set and folds \(1, ..., (i - 1)\) folds are the training set.

Methods (by class)

data.frame: Rows (observations) are split into folds. The ordering of the data frame is assmed to be meaningful.

grouped_df: The groups are split into folds. The ordering of the groups is assumed to be meaningful. No stratification option is provided since the cases in which stratification can usually be rewritten to operate on the groups rather than within groups.

References

Scikit-learn v. 0.18.1. Cross-validation: evaluating estimator performance. http://scikit-learn.org/stable/modules/cross_validation.html#cross-validation.