Generate permutation subsets

Generate R replicates with permutations of the rows or groups of a data frame.

permute(data, ...)

# S3 method for data.frame
permute(data, R = 1L, ...)

# S3 method for grouped_df
permute(data, R = 1L, stratify = FALSE, ...)

Arguments

data

A data frame

...

Arguments passed to methods

R

The number of permutations to generate.

stratify

If TRUE, the permute observations within each group. Otherwise permute the groups themselves.

Value

A data frame with R rows and the following columns:

train

A list of resample objects. Training sets.

test

A list of resample objects. Test sets.

.id

An integer vector of identifiers

Details

If R is greater than the total number of permutations, only the total number of permutations is returned. Since the total number of permutations is factorial(n), this will only happen if n is small.

Methods (by class)

data.frame: Permute rows of a data frame.

grouped_df: Permutes rows within each group in a grouped data frame if stratify = TRUE. Otherwise, permute the groups.