Generate subsamples from indices

Generate subsamples from manually specified indices of observations in each subsample.

resample_df(data, ...)

# S3 method for data.frame
resample_df(data, samples, ...)

# S3 method for grouped_df
resample_df(data, samples, ...)

Arguments

data

A data frame

...

Arguments passed to methods

samples

A list of integer vectors or an integer vector, with the indices in each subsample.

Value

A data frame with rows for each subsample and the following columns:

train

A list of resample objects. Training sets.

test

A list of resample objects. Test sets.

.id

An integer vector of identifiers

Methods (by class)

data.frame: Generate subsamples of rows of data frames.

grouped_df: Generate subsamples of groups of grouped data frames.

Examples

library("dplyr")
# specify subsamples by a list of indices
resample_df(mtcars, samples = c(1:5, 5:10, 15:20))

#> # A tibble: 1 × 2
#>           sample   .id
#>           <list> <int>
#> 1 <S3: resample>     1

resample_df(mtcars, samples = c(1:10, 1:15, 5:30))

#> # A tibble: 1 × 2
#>           sample   .id
#>           <list> <int>
#> 1 <S3: resample>     1

# with grouped data frames, indices refer to groups
resample_df(group_by(mtcars, mpg), samples = list(c(1:2)))$sample

#> [[1]]
#> <resample [3 x 11]> 15, 16, 24
#>